Check that stubs implement intended interface.
Recall that `weaver generate` generates three kinds of stubs for a
component: a local stub, a client stub, and a server stub. The local and
client stubs are supposed to implement the component interface, and the
server stub is supposed to implement the `codegen.Server` interface.

This commit generates code to check that these stubs implement the
intended interfaces. Without these checks, if you add a method to a
component interface and forget to re-run `weaver generate`, the app will
panic at runtime. With these checks, the app will fail to build, which
is preferable.